Upstream-Status: Inappropriate [OE specific]
commonio.c: fix unexpected open failure in chroot environment

When using commands with '-R <newroot>' option in our pseudo environment,
we would usually get the 'Pemission Denied' error. This patch serves as
a workaround to this problem.

Note that this patch doesn't change the logic in the code, it just expands
the codes.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Nobuhiro Iwamatsu <nobuhiro.iwamatsu@miraclelinux.com>